3я часть димпломного проекта Гарейшина Дарья (исправление)
