--- src/test/java/Api/UserData.java
@@ -1,28 +1,0 @@
-package Api;
-
-public class UserData {
-
-    private String success;
-    private String accessToken;
-    private String refreshToken;
-    private UserPersonalData user;
-
-    public String getSuccess() {
-        return success;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
-
-    public String getRefreshToken() {
-        return refreshToken;
-    }
-
-    public UserPersonalData getUser() {
-        return user;
-    }
-
-
-
-}
