--- src/test/java/Api/UserPersonalData.java
@@ -1,15 +1,0 @@
-package Api;
-
-public class UserPersonalData {
-
-    private String email;
-    private String name;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
