+++ src/test/java/PersonalActionsTest.java
@@ -1,27 +1,22 @@
-import Api.UserApi;
-import Api.UserData;
-import PageObject.LoginPage;
-import PageObject.MainPage;
-import PageObject.PersonalAccountPage;
+import api.UserApi;
+import api.UserData;
+import pageobject.LoginPage;
+import pageobject.MainPage;
+import pageobject.PersonalAccountPage;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-import static PageObject.Constants.*;
+import static pageobject.Constants.*;
 import static org.openqa.selenium.By.xpath;
 
-public class PersonalActionsTest {
-
-    private WebDriver driver;
+public class PersonalActionsTest extends BaseTest {
 
     private String token;
 
@@ -30,9 +25,10 @@ public class PersonalActionsTest {
         UserApi user = new UserApi(email, password, name);
         user.creationOfUser(user);
         token = user.authorizationOfUser(user).as(UserData.class).getAccessToken();
-        ChromeOptions options = new ChromeOptions();
-        driver = new ChromeDriver(options);
-        driver.get(LOGIN_PAGE_URL);
+        System.out.println(token);
+        driver.get(MAIN_PAGE_URL);
+        MainPage objMainPage = new MainPage(driver);
+        objMainPage.clickOnPersonalAccountButton();
         LoginPage objLoginPage = new LoginPage(driver);
         objLoginPage.inputLogin(email);
         objLoginPage.inputPassword(password);
