+++ src/test/java/RegistrationTest.java
@@ -1,26 +1,22 @@
-import Api.UserApi;
-import Api.UserData;
-import PageObject.LoginPage;
-import PageObject.RegisterPage;
+import api.UserApi;
+import api.UserData;
+import pageobject.LoginPage;
+import pageobject.RegisterPage;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
-import static PageObject.Constants.*;
+import static pageobject.Constants.*;
 import static org.openqa.selenium.By.xpath;
 
-public class RegistrationTest {
+public class RegistrationTest extends BaseTest {
 
-    private WebDriver driver;
 
     private String token;
 
@@ -28,9 +24,6 @@ public class RegistrationTest {
 
     @Before
     public void setUp(){
-
-        ChromeOptions options = new ChromeOptions();
-        driver = new ChromeDriver(options);
         driver.get(REGISTER_PAGE_URL);
     }
