+++ src/test/java/api/UserPersonalData.java
@@ -1,0 +1,15 @@
+package api;
+
+public class UserPersonalData {
+
+    private String email;
+    private String name;
+
+    public String getEmail() {
+        return email;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
