+++ src/test/java/pageobject/Constants.java
@@ -1,0 +1,21 @@
+package pageobject;
+
+public class Constants {
+
+    public static final String MAIN_PAGE_URL = "https://stellarburgers.nomoreparties.site";
+
+    public static final String MAIN_PAGE_AUTHORIZED_USER_URL = "https://stellarburgers.nomoreparties.site/";
+    public static final String LOGIN_PAGE_URL = "https://stellarburgers.nomoreparties.site/login";
+    public static final String REGISTER_PAGE_URL = "https://stellarburgers.nomoreparties.site/register";
+
+
+    public static final String email = "agatha5@gmail.com";
+    public static final String password = "agatha5";
+
+    public static final String wrongPassword = "aga";
+
+    public static final String name = "Agatha5";
+
+
+
+}
